Direct initial request via the web server
