+++ Web Flow With External Server Posting Back/force-app/main/objects/WebFlowClient__mdt/fields/PublicServerRoot__c.field-meta.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<CustomField xmlns="http://soap.sforce.com/2006/04/metadata">
+    <fullName>PublicServerRoot__c</fullName>
+    <externalId>false</externalId>
+    <label>public web server for offload demo</label>
+    <description>If demonstrating offloading of secrets to the public server enter the address here. Example https://my-domain.com/</description>
+    <length>80</length>
+    <required>false</required>
+    <trackTrending>false</trackTrending>
+    <type>Text</type>
+    <unique>false</unique>
+</CustomField>
