+++ Web Flow With External Server Posting Back/webserver/package.json
@@ -4,11 +4,11 @@
   "description": "Demonstrate oAuth web flow using a NodeJS server",
   "main": "index.js",
   "scripts": {
-    "build": "npx webpack --config wenpack.conf.js",
+    "build": "npx webpack",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
-  "license": "UNLICENSED",
+  "author": "Richard Corfield",
+  "license": "BSD-3-Clause",
   "dependencies": {
     "express": "^4.18.2"
   },
